add members from community when creating activity
