+++ web/src/components/Page/Page.css
@@ -1,9 +1,9 @@
-.ActivitiesPage {
+.Page {
   padding: 15px;
   margin-bottom: 200px;
 }
 
-.ActivitiesPage__Title {
+.Page__Title {
   margin-bottom: 0;
   padding-left: 10px;
   color: #767e84;
@@ -11,11 +11,11 @@
 }
 
 @media only screen and (min-width: 1024px) {
-  .ActivitiesPage__Content {
+  .Page__Content {
     display: flex;
   }
 
-  .ActivitiesPage__Title {
+  .Page__Title {
     font-size: 1.8rem;
   }
 }
