+++ web/src/pages/AppLayout/AppLayout.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { ToastContainer } from 'react-toastify';
 import { Route, Switch, Redirect, NavLink } from 'react-router-dom';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faUsers, faHome, faBuilding } from '@fortawesome/free-solid-svg-icons'
 
 import {
