--- web/src/pages/CommunitiesPage/CommunitiesPage.css
@@ -1,21 +1,0 @@
-.CommunitiesPage {
-  padding: 15px;
-  margin-bottom: 200px;
-}
-
-.CommunitiesPage__Title {
-  margin-bottom: 0;
-  padding-left: 10px;
-  color: #767e84;
-  font-size: 1.5rem;
-}
-
-@media only screen and (min-width: 1024px) {
-  .CommunitiesPage__Content {
-    display: flex;
-  }
-
-  .CommunitiesPage__Title {
-    font-size: 1.8rem;
-  }
-}
