--- web/src/pages/UsersPage/UsersPage.css
@@ -1,21 +1,0 @@
-.UsersPage {
-  padding: 15px;
-  margin-bottom: 200px;
-}
-
-.UsersPage__Title {
-  margin-bottom: 0;
-  padding-left: 10px;
-  color: #767e84;
-  font-size: 1.5rem;
-}
-
-@media only screen and (min-width: 1024px) {
-  .UsersPage__Content {
-    display: flex;
-  }
-
-  .UsersPage__Title {
-    font-size: 1.8rem;
-  }
-}
