+++ web/src/redux/Communities/CommunitiesActionTypes.js
@@ -1,0 +1,3 @@
+export const SET_COMMUNITIES = 'SET_COMMUNITIES';
+export const SET_COMMUNITIES_LOADING_STATE = 'SET_COMMUNITIES_LOADING_STATE';
+export const SET_COMMUNITIES_ERROR = 'SET_COMMUNITIES_ERROR';
